Core command to create img
with a lot of refactor, this needs to be split up into a couple of MPs for easier review

+++ diskimage/image.go
@@ -28,13 +28,36 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 
-	"launchpad.net/goget-ubuntu-touch/ubuntu-emulator/sysutils"
+	"launchpad.net/goget-ubuntu-touch/sysutils"
 )
 
+type imageLabel string
+type directory string
+
+const (
+	systemDataLabel imageLabel = "system-data"
+	userDataLabel   imageLabel = "user-data"
+)
+
+const (
+	systemDataDir directory = "system"
+	userDataDir   directory = "user-data"
+)
+
+type partition struct {
+	label imageLabel
+	dir   directory
+	loop  string
+}
+
 type DiskImage struct {
-	path, label, Mountpoint string
-	size                    int64
+	Mountpoint string
+	label      string
+	path       string
+	size       int64
+	parts      []partition
 }
 
 //New creates a new DiskImage
@@ -91,6 +114,32 @@ func (img *DiskImage) Copy(dst string) (err error) {
 	return nil
 }
 
+// User returns the user path
+func (img *DiskImage) User() (string, error) {
+	if img.parts == nil {
+		return "", errors.New("img is not setup with partitions")
+	}
+
+	if img.Mountpoint == "" {
+		return "", errors.New("img not mounted")
+	}
+
+	return filepath.Join(img.Mountpoint, string(userDataDir)), nil
+}
+
+//System returns the system path
+func (img *DiskImage) System() (string, error) {
+	if img.parts == nil {
+		return "", errors.New("img is not setup with partitions")
+	}
+
+	if img.Mountpoint == "" {
+		return "", errors.New("img not mounted")
+	}
+
+	return filepath.Join(img.Mountpoint, string(systemDataDir)), nil
+}
+
 //Mount the DiskImage
 func (img *DiskImage) Mount() (err error) {
 	img.Mountpoint, err = ioutil.TempDir(os.TempDir(), "ubuntu-system")
@@ -103,8 +152,29 @@ func (img *DiskImage) Mount() (err error) {
 			os.Remove(img.Mountpoint)
 		}
 	}()
+	if img.parts == nil {
+		return img.emulatorMount()
+	}
+
+	return img.coreMount()
+}
+
+func (img *DiskImage) coreMount() (err error) {
+	for _, part := range img.parts {
+		mountpoint := filepath.Join(img.Mountpoint, string(part.dir))
+		if err := os.MkdirAll(mountpoint, 0755); err != nil {
+			return err
+		}
+		if out, err := exec.Command("mount", filepath.Join("/dev/mapper", part.loop), mountpoint).CombinedOutput(); err != nil {
+			return fmt.Errorf("unable to dir to create system image: %s", out)
+		}
+	}
+	return nil
+}
+
+func (img *DiskImage) emulatorMount() (err error) {
 	if out, err := exec.Command("mount", img.path, img.Mountpoint).CombinedOutput(); err != nil {
-		return errors.New(fmt.Sprintf("Unable to mount temp dir to create system image %s", out))
+		return fmt.Errorf("unable to mount temp dir to create system image: %s", out)
 	}
 	return nil
 }
@@ -116,13 +186,41 @@ func (img *DiskImage) Unmount() error {
 	}
 	defer os.Remove(img.Mountpoint)
 	if out, err := exec.Command("sync").CombinedOutput(); err != nil {
-		exec.Command("umount", img.Mountpoint).Run()
 		return errors.New(fmt.Sprintf("Failed to sync filesystems before unmounting: %s", out))
 	}
+
+	if img.parts == nil {
+		if err := img.emulatorUnmount(); err != nil {
+			return err
+		}
+	} else {
+		if err := img.coreUnmount(); err != nil {
+			return err
+		}
+	}
+
+	img.Mountpoint = ""
+
+	return nil
+}
+
+func (img *DiskImage) coreUnmount() (err error) {
+	for _, part := range img.parts {
+		mountpoint := filepath.Join(img.Mountpoint, string(part.dir))
+		if out, err := exec.Command("umount", mountpoint).CombinedOutput(); err != nil {
+			return fmt.Errorf("unable to dir to create system image: %s", out)
+		}
+		if err := os.Remove(mountpoint); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (img *DiskImage) emulatorUnmount() (err error) {
 	if err := exec.Command("umount", img.Mountpoint).Run(); err != nil {
 		return errors.New("Failed to unmount temp dir where system image was created")
 	}
-	img.Mountpoint = ""
 	return nil
 }
 
@@ -146,15 +244,100 @@ func (img *DiskImage) Provision(tarList []string) error {
 	return nil
 }
 
-//Create returns a ext4 partition for a given file
-func (img DiskImage) CreateExt4() error {
+//Partition creates a partitioned image from an img
+func (img *DiskImage) Partition() error {
 	if err := sysutils.CreateEmptyFile(img.path, img.size); err != nil {
 		return err
 	}
-	return exec.Command("mkfs.ext4", "-F", "-L", img.label, img.path).Run()
+
+	partedCmd := exec.Command("parted", img.path)
+	stdin, err := partedCmd.StdinPipe()
+	if err != nil {
+		return err
+	}
+
+	stdin.Write([]byte("mklabel msdos\n"))
+	stdin.Write([]byte("mkpart primary ext4 2048s 4G\n"))
+	stdin.Write([]byte("mkpart primary ext4 4G -1s\n"))
+	stdin.Write([]byte("set 1 boot on\n"))
+	stdin.Write([]byte("print\n"))
+	stdin.Write([]byte("quit\n"))
+
+	return partedCmd.Run()
 }
 
-//Create returns a vfat partition for a given file
+//MapPartitions creates loop devices for the partitions
+func (img *DiskImage) MapPartitions() error {
+	kpartxCmd := exec.Command("kpartx", "-avs", img.path)
+	stdout, err := kpartxCmd.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	if err := kpartxCmd.Start(); err != nil {
+		return err
+	}
+
+	loops := make([]string, 0, 2)
+	scanner := bufio.NewScanner(stdout)
+	for scanner.Scan() {
+		fields := strings.Fields(scanner.Text())
+		if len(fields) > 2 {
+			loops = append(loops, fields[2])
+		} else {
+			return errors.New("issues while determining drive mappings")
+		}
+	}
+	if err := scanner.Err(); err != nil {
+		return err
+	}
+
+	if len(loops) != 2 {
+		return errors.New("more partitions then expected while creating loop mapping")
+	}
+
+	img.parts = []partition{
+		partition{label: systemDataLabel, dir: systemDataDir, loop: loops[0]},
+		partition{label: userDataLabel, dir: userDataDir, loop: loops[1]},
+	}
+
+	if err := kpartxCmd.Wait(); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+//UnMapPartitions destroys loop devices for the partitions
+func (img *DiskImage) UnMapPartitions() error {
+	for _, part := range img.parts {
+		if err := exec.Command("dmsetup", "clear", part.loop).Run(); err != nil {
+			return err
+		}
+	}
+	return exec.Command("kpartx", "-d", img.path).Run()
+}
+
+//CreateExt4 returns a ext4 partition for a given file
+func (img DiskImage) CreateExt4() error {
+	if img.parts == nil {
+		if err := sysutils.CreateEmptyFile(img.path, img.size); err != nil {
+			return err
+		}
+		return exec.Command("mkfs.ext4", "-F", "-L", img.label, img.path).Run()
+	}
+
+	for _, part := range img.parts {
+		dev := filepath.Join("/dev/mapper", part.loop)
+		if err := exec.Command("mkfs.ext4", "-F", "-L", string(part.label), dev).Run(); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+//CreateVFat returns a vfat partition for a given file
 func (img DiskImage) CreateVFat() error {
 	if err := sysutils.CreateEmptyFile(img.path, img.size); err != nil {
 		return err
